Add test to cover more parts of Maintenance.models
- Also moved tests for find_all_files from views to models.
- Minor cleanup of unused variables.
